requirements.txt
Flask
aiogram
python-docx
